🐛 Fix clash between k8s svc discovery and rig env var
Kubernetes service discovery creates an environment variable called
`<service-name>-PORT` for each service in the namespace where a pod
runs. The main service in the chart used the fullname helper as name.
When using rig as release name this resulted in the service being called
`rig`. This then means that each pod in the namespace would get an
environment variable called `RIG_PORT`. This clashes with the
environment override of `port` in the rig config.

This fix changes the service name by adding `-svc` as suffix. The
service discovery variable will hence be `RIG_SVC_PORT` and no longer
cause a clash.