CLI: Add --remote flag to create-build and deploy (#121)
The flag specifies if an image should not be searched for locally.
Sometimes you have a remote image also pulled locally. In that case,
you might not want to refer to the local one (which using a dev-registry
would unnecessarily be pushed into the dev-reg if considered a local image).