🔥 Remove unused files
As a startup we need to eventually find a way to make money to create a
sustainable business. Coming up with a business model to open-source
projects are always hard. We don't want to offer a managed solution
which is the most common strategy when running full open-source. This
has led us to decide to only be partially open-source for now.

🌍  What will remain open-source:

- `rig-operator`. Our abstractions on top of kubernetes
- `rig` The CLI for interacting with the `rig-platform` and it's resources.
- Helm charts for installing `rig-operator` and `rig-platform`.
- The platform protobuf interfaces (allows for easy API-client generation)

🔒 What will be closed-source (at least for now):

- The `rig-platform` which consists of the dashboard and a better
  developer experience with regards to working with the abstractions
  provided by `rig-operator`.

This split will allow us to create a paid product which we can go to
market with together with support offerings, and at the same time we
will keep our core kubernetes abstractions open-source. We think that
the `rig-operator` by it self makes a lot of sense, as it standardises
what resources are created for an application and how they are
configured. It also provides security for our customers in the event
that they should choose not to be `rig-platform` customers any more, they
will still keep the abstractions and all of the systems they deployed
through the `rig-platform`.

We're looking forward to shipping updates soon, and getting feedback on
the new version. 🦾  🙏