✨ Watch for config changes
The operator now watches configmaps and secrets for changes in order to
determine wether we need to reconcile a capsule.

Eg. when we depend on a secret or a configmap for environment variables
or configfiles, we want to update an annotation on the deployment when
these change so that we roll the deployment.

- ✨ Introduce a hash function which we use for calculating a hash of the
  current configuration eg. contents of configmaps and secrets we depend
  upon.
- 🐛 Fix a few bugs in `upsertIfNewer`
- ✅ Disable validation integration test tracked in #226
- 🔊 Improve logging