:bug: Ensure leader election uses same namespace as operator
Leader election namespace was hardcoded to "rig-system". When deploying
in other namespaces the granted role didnt allow for access to leases in
rig-system but instead in the namespace where it was deployed.

This change uses the downward API to make an environment variable
available in the pod for what namespace we are running in. This can then
be used to ensure that we use lease resources in the same namespace as
where the operator is running.