:recycle: Refactored how we handle templating in plugins (#647)
Part of the refactor is to also handle cases where we need to
do templating in a series of steps. E.g. objectTemplate can use
capsule and currentObject as template values, if the name of currentObject
is templated on the capsule, then you first need to template the config
using the capsule, then fetch the currentObject, then template again.